pyproject.toml
[project]
name = "idt-calculator"
version = "0.1.0"
description = "An ACES Input Device Transform (IDT) calculator"
readme = "README.md"
requires-python = ">=3.10,< 3.14"
authors = [
    { name = "Alex Forsythe", email = "aforsythe@oscars.org" },
    { name = "Thomas Mansencal",  email = "thomas.mansencal@gmail.com" },
    { name = "Adam Davis",  email = "adamdavis@netflix.com" },
]
maintainers = [
    { name = "Alex Forsythe", email = "aforsythe@oscars.org" },
    { name = "Thomas Mansencal",  email = "thomas.mansencal@gmail.com" },
    { name = "Adam Davis",  email = "adamdavis@netflix.com" },
]
license =  { text = "License Terms for Academy Color Encoding System Components" }
keywords = [
    "aces",
    "ampas",
    "color",
    "color-science",
    "color-space",
    "color-spaces",
    "colorspace",
    "colorspaces",
    "colour",
    "colour-science",
    "colour-space",
    "colour-spaces",
    "colourspace",
    "colourspaces",
    "dash",
    "docker",
    "python"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development"
]

dependencies = [
    "colour-datasets>=0.2.5",
    "colour-checker-detection>=0.2.1",
    "colour_science>=0.4.5",
    "dash",
    "dash-bootstrap-components",
    "dash-renderer",
    "dash-uploader",
    "gunicorn",
    "imageio>=2,<3",
    "pandas>=2,<3",
    "jsonpickle>=2,<3",
    "matplotlib>=3.7",
    "networkx>=3,<4",
    "numpy>=1.24,<3",
    "packaging<= 21.3",  # Later versions currently break with Dash Uploader.
    "scipy>=1.10,<2",
    "plotly",
    "typing-extensions>=4,<5",
    "xxhash>=3,<4",
]

[project.urls]
Homepage = "https://www.oscars.org/science-technology/sci-tech-projects/aces"
Repository = "https://github.com/ampas/idt-calculator"
Issues = "https://github.com/ampas/idt-calculator/issues"
Changelog = "https://github.com/ampas/idt-calculator/releases"

[tool.uv]
package = true
dev-dependencies = [
    "coverage",
    "coveralls",
    "hatch",
    "invoke",
    "jupyter",
    "pre-commit",
    "pyright",
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "toml",
    "twine",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [ "aces" ]

[tool.codespell]
ignore-words-list = "rIn"

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
split_on_trailing_comma = true
use_parentheses = true

[tool.pyright]
reportMissingImports = false
reportMissingModuleSource = false
reportUnboundVariable = false
reportUnnecessaryCast = true
reportUnnecessaryTypeIgnorComment = true
reportUnsupportedDunderAll = false
reportUnusedExpression = false

[tool.pytest.ini_options]
addopts = "-n auto --dist=loadscope --durations=5"
filterwarnings = [
    "ignore::RuntimeWarning",
    "ignore::pytest.PytestCollectionWarning",
]

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["ALL"]
ignore = [
    "C", # Pylint - Convention
    "C90", # mccabe
    "COM", # flake8-commas
    "ERA", # eradicate
    "FBT", # flake8-boolean-trap
    "FIX", # flake8-fixme
    "PT", # flake8-pytest-style
    "PTH", # flake8-use-pathlib [Enable]
    "TD", # flake8-todos
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed in `**kwargs`
    "D200", # One-line docstring should fit on one line
    "D202", # No blank lines allowed after function docstring
    "D205", # 1 blank line required between summary line and description
    "D301", # Use `r"""` if any backslashes in a docstring
    "D400", # First line should end with a period
    "I001", # Import block is un-sorted or un-formatted
    "N801", # Class name `.*` should use CapWords convention
    "N802", # Function name `.*` should be lowercase
    "N803", # Argument name `.*` should be lowercase
    "N806", # Variable `.*` in function should be lowercase
    "N813", # Camelcase `.*` imported as lowercase `.*`
    "N815", # Variable `.*` in class scope should not be mixedCase
    "N816", # Variable `.*` in global scope should not be mixedCase
    "NPY002", # Replace legacy `np.random.random` call with `np.random.Generator`
    "PGH003", # Use specific rule codes when ignoring type issues
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison, consider replacing `.*` with a constant variable
    "PYI036", # Star-args in `.*` should be annotated with `object`
    "PYI051", # `Literal[".*"]` is redundant in a union with `str`
    "PYI056", # Calling `.append()` on `__all__` may not be supported by all type checkers (use `+=` instead)
    "RUF022", # [*] `__all__` is not sorted
    "TRY003", # Avoid specifying long messages outside the exception class
    "UP038", #  Use `X | Y` in `isinstance` call instead of `(X, Y)`
]
typing-modules = ["colour.hints"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.per-file-ignores]
"__init__.py" = ["D104"]
"apps/*" = ["ANN"]
"docs/*" = ["INP"]
"tasks.py" = ["INP"]
"test_*" = ["S101"]

[tool.ruff.format]
docstring-code-format = true
